cmd/anubis: amend generic browser challenge to "Gecko"
I'm gonna be honest, this is an extreme galaxy brain strategy and I'm
not entirely sure if this will pan out. However I got the idea when
reading [a community post][0]

If this works, that would be so much funnier than just using "Mozilla"
in the rules. I think that this could greatly backfire though, which is
why I'm making a pull request and opening this for feedback from the
community.

It would be absolutely hilarious if this works though.

[0]: https://github.com/TecharoHQ/anubis/discussions/68#discussioncomment-12583134